fix: open design of plugin instantiation
